% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g11_beta.R
\name{g11_beta}
\alias{g11_beta}
\title{Least Square}
\usage{
g11_beta(response, covariates)
}
\arguments{
\item{response}{A \code{vector} with the values for the dependent variable (also called outcome).}

\item{covariates}{A \code{matrix} with the values for the independent variable (also called predictors, or explanatory variable).}
}
\value{
A \code{numeric}.
}
\description{
Function that provides the least square estimator (beta_hat)
}
\examples{
Using data(hubble) from libary(gamair)
g11_lm(hubble$y, hubble$x)
}
\author{
Group11
}
