% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g11_mspe.R
\name{g11_mspe}
\alias{g11_mspe}
\title{}
\usage{
g11_mspe(response, covariates)
}
\arguments{
\item{response}{A \code{vector} with the values for the dependent variable (also called outcome).}

\item{covariates}{A \code{matrix} with the values for the independent variable (also called predictors, or explanatory variable).}

\item{Beta}{A \code{matrix} that can be calculated using \code{g11_lm} function.!}

\item{alpha}{A \code{numeric} (double) that sets the alpha coefficient to be used. Has to be between 0 and 1.}

\item{method}{A \code{string} that defines the method used. Options are "Asymptotic" and "Bootstrap", accepts minor misspellings with a warning - which can be both good and bad.}
}
\value{
A \code{list} containing the following attributes:
\describe{
     \item{beta}{Estimated coefficients, Linear Regression Model.}
     \item{sigma2}{explanation}
     \item{variance_beta}{explanation}
     \item{ci}{explanation}
}
}
\description{
Function that
}
\examples{
Using data(hubble) from libary(gamair)
g11_lm(hubble$y, hubble$x, alpha = 0.01, method = "Bootstrap")
g11_lm(hubble$y, hubble$x, method = "Asymptotic")

From here, is the ACTUAL function:
}
\author{
Group11
}
